+++ remote/src/main/java/com/kdjj/remote/datasource/RecipeRemoteDataSourceImpl.kt
@@ -4,7 +4,7 @@ import com.kdjj.data.common.errorMap
 import com.kdjj.data.datasource.RecipeRemoteDataSource
 import com.kdjj.domain.model.Recipe
 import com.kdjj.remote.common.fireStoreExceptionToDomain
-import com.kdjj.remote.dao.RemoteRecipeService
+import com.kdjj.remote.service.RemoteRecipeService
 import javax.inject.Inject
 
 internal class RecipeRemoteDataSourceImpl @Inject constructor(
