+++ remote/src/main/java/com/kdjj/remote/di/RemoteRecipeListModule.kt
@@ -1,8 +1,8 @@
 package com.kdjj.remote.di
 
 import com.kdjj.data.datasource.RecipeListRemoteDataSource
-import com.kdjj.remote.dao.RecipeListServiceImpl
-import com.kdjj.remote.dao.RemoteRecipeListService
+import com.kdjj.remote.service.RecipeListServiceImpl
+import com.kdjj.remote.service.RemoteRecipeListService
 import com.kdjj.remote.datasource.RecipeListRemoteDataSourceImpl
 import dagger.Binds
 import dagger.Module
