+++ remote/src/main/java/com/kdjj/remote/service/FirebaseStorageServiceImpl.kt
@@ -1,4 +1,4 @@
-package com.kdjj.remote.dao
+package com.kdjj.remote.service
 
 import android.net.Uri
 import com.google.firebase.storage.StorageReference
