+++ remote/src/main/java/com/kdjj/remote/service/RecipeServiceImpl.kt
@@ -1,21 +1,16 @@
-package com.kdjj.remote.dao
+package com.kdjj.remote.service
 
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.Source
 import com.google.firebase.firestore.ktx.toObject
 import com.kdjj.domain.model.Recipe
-import com.kdjj.domain.model.exception.ApiException
-import com.kdjj.domain.model.exception.NetworkException
 import com.kdjj.remote.dto.RecipeDto
 import com.kdjj.remote.dto.toDomain
 import com.kdjj.remote.dto.toDto
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.coroutines.tasks.await
 import kotlinx.coroutines.withContext
-import java.util.concurrent.ThreadPoolExecutor
 import javax.inject.Inject
-import kotlin.coroutines.resumeWithException
 
 internal class RecipeServiceImpl @Inject constructor(
     private val firestore: FirebaseFirestore
