+++ remote/src/test/java/com/kdjj/remote/recipe/RecipeImageRemoteDataSourceImplTest.kt
@@ -1,6 +1,6 @@
 package com.kdjj.remote.recipe
 
-import com.kdjj.remote.dao.FirebaseStorageServiceImpl
+import com.kdjj.remote.service.FirebaseStorageServiceImpl
 import com.kdjj.remote.datasource.RecipeImageRemoteDataSourceImpl
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
